# Perf: accélération du chargement des dossiers visibles