--- a/db/migrate/20231110135532_alter_dossiers_for_procedure_preview_not_nullable.rb
+++ b/db/migrate/20231110135532_alter_dossiers_for_procedure_preview_not_nullable.rb
@@ -1,0 +1,5 @@
+class AlterDossiersForProcedurePreviewNotNullable < ActiveRecord::Migration[7.0]
+  def change
+    add_check_constraint :dossiers, "for_procedure_preview IS NOT NULL", name: "dossiers_for_procedure_preview_null", validate: false
+  end
+end