--- a/db/migrate/20231110135533_validate_alter_dossiers_for_procedure_preview_not_nullable.rb
+++ b/db/migrate/20231110135533_validate_alter_dossiers_for_procedure_preview_not_nullable.rb
@@ -1,0 +1,7 @@
+class ValidateAlterDossiersForProcedurePreviewNotNullable < ActiveRecord::Migration[7.0]
+  def change
+    validate_check_constraint :dossiers, name: "dossiers_for_procedure_preview_null"
+    change_column_null :dossiers, :for_procedure_preview, false, false
+    remove_check_constraint :dossiers, name: "dossiers_for_procedure_preview_null"
+  end
+end