Les `OR` dans les requêtes PG sont généralement une source de ralentissement car elles freinent certaines optimisations.

En supprimant `OR for_procedure_preview IS NULL` du scope `Dossiers.visible_by_user` on améliore le scope jusqu'à ~30% pour récupérer des dossiers (c'est à peu près ISO pour un count). Sur certaines démarches ça prend plusieurs centaines de ms

La colonne n'était pas non nullable pour les besoins de la création mais ça avait été rattrapé ici ,et d'ailleurs on a aucun dossier avec une valeur NULL.
[`lib/tasks/deployment/20220427195148_set_default_for_procedure_preview_on_dossiers.rake`](https://github.com/demarches-simplifiees/demarches-simplifiees.fr/commit/27f9a820cc59652c5f9095b8e2affdcf33f05ded#diff-7202c64222deec0c5957bb82c4e6d655be822b5f2db661eeb1cfd09402812a6c)

Cette PR enforce donc la contrainte non nullable, et optimise le scope.


```ruby
  task dossiers_visible_by_user: :environment do
    require 'benchmark/ips'
    procedure = Procedure.find(62746)

    Benchmark.ips do |x|
      x.report("fast") { procedure.dossiers.visible_by_user.all }
      x.report("slow") { procedure.dossiers.visible_by_user_slow.all }
      x.compare!
    end
  end


Warming up --------------------------------------
                fast     1.738k i/100ms
                slow     1.341k i/100ms
Calculating -------------------------------------
                fast     17.214k (± 3.0%) i/s -     86.900k in   5.052921s
                slow     12.926k (± 3.6%) i/s -     65.709k in   5.090313s

Comparison:
                fast:    17213.9 i/s
                slow:    12926.1 i/s - 1.33x  slower
```